[FIX] test_discuss_full_enterprise: fix crash 'hr.employee' missing
Before this commit, recently added unit test "can handle command and
disable mentions in AI composer" crashed on runbot with the following
error message:

```
cannot find a definition for model "hr.employee": could not get model
from server environment (did you forget to use `defineModels()?`)
```

Somehow the mock models in this test suite is aware of model
"hr.employee". While this model exists with this module installed,
the hoot suite in `test_discuss_full_enterprise` only imports
mail models which hasn't `hr.employee`, thus we'd expect the suite to
not be aware of this model and shouldn't attempt to load it.

While we don't know why the hr.employee model definition is obtained
on runbot on this test suite, functionally it makes sense to have
this model installed. At the same time, we'd expect that the JS model
definition are optional and just there in case we want to "cheat"
model definition from the one fetched.

This commit attempts to fix the issue by defining hr models rather
than mail models in this test suite. Hr models have mail models in
addition to hr models like hr.employee, thus we expect to fix the
crash.

Fixes runbot-error-230073

closes odoo/enterprise#90608

X-original-commit: 43134070162baeedba79832fd215c87c2497a415
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>