[FIX] web_gantt: proper translations for gantt cards
Steps to Reproduce:

- Change the language to something other than English.

- Create a new menu in the Appointment module using Studio.

- Link that menu with the existing model (Resource Time of Detail).

- Save the changes and add the Gantt view

- Save the changes and close Studio.

- View this new menu in the Gantt view

Issue:

- When clicking onto the kanban popover card for any record,
the labels will remain as "Name" , "Start" , and "Stop" in
English, and it will not translate.

Reason:

- [1] changed the rendering of default kanban card
templates from XML to JS. This changed the default template
to be static.

Fix:

- Add a call to the translation core tool "_t" on the three
labels upon their generation in _getKanbanViewParams(). This
ensures that the arch template that is generated is translated
correctly into the user's set language.

- Additionally, add these terms into the .pot file for web_gantt.
This ensures that the proper translations are stored for transifex
and the relevant translation team and that the call to _t will have
a translation to reference in it's call.

[1]: https://github.com/odoo/enterprise/commit/b2e31239041e6b7850888a3ae070c3f8a40ea5ee

opw-4861899

closes odoo/enterprise#93097

X-original-commit: 3cb924a6322d25657f9e4ff058010dd6e866943f
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>