[IMP] ai: reduce the number of API call with a new argument
Purpose
=======
When the LLM is done executing action, there's no way to know it
except if we make one more API call, and see no "call actions".

So, if the LLM execute one action and stop, it will make 2 API calls.

To reduce that number to one, we add a new argument `__end_message`
and tell the LLM to fill that argument when it's done.

This is the equivalent to `AITerminate`, but work with any action
and it's automatic.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>