[FIX] knowledge: fix crash occuring in tours
This commit fixes the template, that could make the SelectMenu
component crash when searchValue is still 'null', since the template
was expecting reading a length from a string.

Without this change, the component being used in the Calendar view
edition within Knowledge was not able to work at all.

The commit also fixes the style of the togglers, because in the same
dialog, a button was being used as the menu toggler for some properties,
instead of using an input. This created inconsistency in the UI, and
made the existing tour fail because the selector was expecting an
empty input.

runbot-error-229945

closes odoo/enterprise#90611

Related: odoo/odoo#219960
Signed-off-by: Pierre Paridans (app) <app@odoo.com>