[IMP] ai: drop `ai.tools` in favor of server actions
Purpose
=======
Remove the `ai.tool` model, and instead use server action marked as
"use with AI" that we introduced in the previous commits.

The idea is to be more flexible in which action we can do.

Adapt the existing `ai.tool` to switch the server actions.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>