[FIX] project_account_budget: update project budget color to be consistent
Issue:
- Color for project budget and total project budgets are inconsistent

Step to reproduce:
- In Settings:
   - activate Project > Timesheet
   - activate Accounting > Budget Management
- Create a project
- Go to Dashboard
- Add a budget of type 'Expense'
- Add a line with a positive amount

Current Behavior:
- Color for the budget is correct
- Color for the total budget is wrong

Expected behavior:
- Color for the budget and total budget are consistent

Cause:
While computing color, expense budget are dealt as if they were negative.
Therefore having a positive amount for expenses result in some problem on the color of the total of different budgets.

Solution:
This commit creates temporary variables to compute progress of the global budget.
Project cards in kanban view use the `total_budget_progress` field, so it also needed to be updated the same way.

opw-4812468

closes odoo/enterprise#97291

X-original-commit: 963ca9c7896972d4692a0ec6d67bbea92fa481ab
Signed-off-by: Corentin Lombard (clom) <clom@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>