[FIX] hr_expense_extract: fix ocr expense creation
How to reproduce the issue:

- Upload a pdf via the OCR with a user without expense right.

On submitting the expense, an access rights error on the write is thrown.

When the expense is created by OCR, on submit, we pass through the action_submit method in hr_expense_extract, which performs validate_ocr, and writes on the expense's extract_state field after the super().action_submit call, which changes the state of the expense from draft to submitted. However, the rule ir_rule_hr_expense_employee_not_draft forbids modifying the expense if it is not in draft state.

opw-4855000

closes odoo/enterprise#90816

X-original-commit: 01f32f2d93b609c122f967a4a716e7d81b982b2d
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>