[FIX] website_sale_{renting,subscription}: locate form relative to button when outside standard hierarchy
## Version
saas-18.4+

## Steps to reproduce
- Open the shop;
- Select any product;
- Open the Editor:
  - Select the product's main image;
  - Change the image width to either `100 percent` or `None`, then save;
- Click on `Add to cart`.

## Issue
Commit odoo/odoo@bbb2d98 introduced a UI update that reorganizes the layout of the product configurator, placing the `Add to cart` button next to the form rather than below it.

Although the button remains inside the form in the original template, using the Editor to adjust the layout can result in the button being saved outside the `<form>` element in the final DOM.

This breaks the logic that relies on `closest('form')` to locate the surrounding form, since the button is no longer a descendant of the form element.

## Solution
Find the first product form relative to the button, since it may be a sibling rather than an ancestor in the DOM.

opw-4942986

closes odoo/enterprise#93306

X-original-commit: 95d632ae774e108b358e8b3a63429afd294dbc52
Related: odoo/odoo#224460
Signed-off-by: Valeriya Chuprina (vchu) <vchu@odoo.com>
Signed-off-by: Pierre Lamotte (pila) <pila@odoo.com>