[FIX] website_appointment: fix traceback on 'meeting' step click
TO REPRODUCE (website must be installed)
============

0. Create a new appointment type, with 'Availability on' 'resources',
at least two resources, and 'Front-end Display' as 'Show Pictures',
and 'Assignment method' as 'Pick Uer/Resource then Time'.

This will ensure the display of the resource selection dedicated
step / screen when booking an appointment ('Meeting' in the progress
bar).

1. On the appointment form, use the 'Share' button.
2. Create a share link with the given appointment type, 'Assign to'
specific Users/Resources, and set to at least two resources in
'Resources'.
3. Copy Link and Close
4. Use the link in the clipboard as a public user
5. Select a resource among those proposed in the 'Meeting' step
6. Go back by clicking the 'Meeting' chevron in the progress bar
7. 403 Error

ISSUE
=====

The filter on resources was not kept in the url when clicking the
'Meeting' step in the progress bar, when they should have been. Hence,
the invitation parameters validation fails.

FIX
===

Add 'filter_resource_ids' in the list of kept parameters.

Task-4897098

closes odoo/enterprise#90802

X-original-commit: 4f7fa37f81f572a024ec186687d5c0072c72584b
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>