[ADD] l10n_om_reports: Tax returns feature
Issue:
This commit adds a missing account.return.type data in the
Oman localization. This is related to the new Accounting Returns #81569

The l10n_om module did not include an account.return.type record,
which prevents the "Returns" button (linked to the action_open_returns)
from appearing in the Tax Report. As a result, users in Oman are unable
to access the Tax Closing form.

Also the l10n_om is a community module, because of that, we need to add
a new enterprise module to add the related data for the tax returns.

The visibility of the button is controlled by the following condition:
[Code](https://github.com/odoo/enterprise/blob/f924f120cbd04002e21c78b081e1c8006b56a4b6/account_reports/models/account_report.py#L1776)

Affected Versions:
18.3 and later

Steps to reproduce:

1.Install the l10n_om module
2.Switch to the Oman company
3.Go to Accounting > Reporting > Tax Report
4.The "Returns" button is not visible

Expected behavior
The "Returns" button should be visible, allowing the user to generate the closing tax form.

Current behavior
The button is hidden due to missing account.return.type data in the Oman localization.

OPW-4943024

closes odoo/enterprise#90943

X-original-commit: b83b1315bd1c2d40f971f0878a900fb9d19a1496
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>
Signed-off-by: Victor Miguel Armenta Carrillo (vmac) <vmac@odoo.com>