[FIX] pos_iot: update customer display using websocket
We fixed the customer display not opening and displaying a traceback when the Iot Box was not reachable,
by calling the action via iot_http, fallbacking on websocket when needed.
We also removed the "customer display with IoT Box" logic from the PoS to the module pos_iot.

Community PR: odoo/odoo#219917

closes odoo/enterprise#90849

X-original-commit: e2a8d5f60149f575b9bdae780957bdec2b5f2be5
Related: odoo/odoo#220234
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>