[FIX] sale_renting_planning: generate shifts when confirming rental order based on uom
Before this commit, when the user creates room product to generate
shifts when a rental order is confirmed with that product, we expect to
generate one shift per quantity ordered if the UoM is Unit instead of
Hours/Days.

This commit improves the shifts generation to take into account the UoM
set on the SOL linked to rental order confirmed. That is, if the UoM set
is Unit then one slot per unit ordered of the product will be generated.
Otherwise, one shift will be generated for the SOL as it is currently
the case.

task-4916537

X-original-commit: cc6ea8be965936fa5d9140b214dbeafdba091762
Part-of: odoo/enterprise#92260
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>