[IMP] web_enterprise: Improve database expiration message
Previously, the database expiration date was set to the same
date as the expiration field, which is now defined as 15 days
after the next invoice date.

Issue: Users tend to wait until the last minute to pay.
This improvement aims to better handle expiration timing and
encourage timely payments.

closes odoo/enterprise#91171

Taskid: 4384877
X-original-commit: f4f75da08925d27e896b8d49ca4ab01455fa9672
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Arnaud Baes (arb) <arb@odoo.com>
Co-authored-by: mva-odoo <mva@odoo.com>