[IMP] l10n_be_reports: 281.50 handle natural person first name
The aim of this commit is to comply with the 281.50 official specs by
handling the first name of natural person and setting it in the xml.

Before this commit:
We don't handle first name and we set the whole partner.name as the
name.

After this commit:
We handle first name and assume it is everything after the last ', '
separation.

closes odoo/enterprise#93099

Task-id: 4919018
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>