[IMP] l10n_br_avatax: correct the NCM code data model
This model was readonly without any views when it was originally
created [1]. We forgot to add the correct constraints, and we duplicated
the code inside the name field to avoid overwriting display_name and
having to add a _rec_name_search.

This commit fixes those issues. First of all, we require both the code
and name field. Second of all, we remove the code from the name field
and modify display_name and _rec_name_search accordingly. To not have
identical display names in a dropdown, we also add a unique constraint
on the combination of name and code. It's possible to have multiple
identical NCM codes with different EX values, but in those cases the
user should give them unique names.

We also take the opportunity to put the NCM form view in a sheet,
looks a bit more modern that way.

[1] https://github.com/odoo/enterprise/pull/43797

task-4998498

closes odoo/enterprise#91750

Related: odoo/upgrade#8178
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>