[FIX] sign: fallback to false in computing share_link
Commit [1] introduced a new field share_link in sign_request that
gets computed for requests that are in the 'shared' state. However this
compute method fails for requests not in the 'shared' state leading to
a traceback error. This commit fixes it by setting the default as False
for the sign.request records that do not have state='shared'

[1]: https://github.com/odoo/enterprise/commit/58425a0022c79f2c45f23fdd5a5476d8c20a887e

opw-4864159

closes odoo/enterprise#90841

X-original-commit: add685279ed434cfc3330611d1ad1ecc5ba5406c
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Yatharth Vyas (yavy) <yavy@odoo.com>