[FIX] sale_subscription: ignore subs from archived companies
Currently, when a company which has recurring orders (subscriptions)
active is archived, the cron still periodically invoices them as usual
but the user is not able to see it because they don't show up in the
views. To avoid having to unarchive the comany and check every
subscription one by one, this PR aims at making the cron ignore any
subscription tied to an archived company, as intended. The only modified
cron is the one that creates invoices since the other one of interest
(the one that sends reminders) operates only on invoiced orders, which
we avoid having with this change.
Also added a test to ensure that if the company related to a subscription
sale order is archived, the cron won't invoice it.

closes odoo/enterprise#90746

Task: 4947107
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>