[PERF] sale_commission: create TEMP table for invoice rules
Before this commit, the invoices rules were the slowest cte used in the achievement and commission reports.
CTE's can't be indexed and when they have a few thousands of rows, they can become quite slow. This commits create a temporary table indexed.
It allows to speed up the join over invoice_rules to use the index. On a production database where the amount of invoice rules is equal to
~ 12000 rows, we could go from 20s to display the commission of a saleperson to 16s to 17s.

Moreover, we could decrease the time to ~12s by disabling the JIT
compilation. Indeed, the JIT would try to do some optimizations that would be wrong
in the context of these requests. By disabling the JIT, we gain the
computation time at the cost of some optimizations. Our tests showed a
~5s gain by disabling the JIT.

task-4783929

closes odoo/enterprise#91183

X-original-commit: 6d1b4d016a6cdffce8921e72cfbb146002c3751d
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>