[FIX] ai_fields: fix insertion of fields and records in prompts
Purpose:
--------
Currently records and fields inserted in AI prompts are badly formatted,
making it sometimes impossible to delete an inserted field or to place
the cursor after a field or a record.
This commit fixes their format:
For standalone fields, we do not insert divs inside a span anymore
For grouped fields, the container is now a div instead of a t (which
was set as inlined by the qweb plugin while the group shouldn't)
For records, the container is now a span instead of a div, since we want
the records to be inlined.
Also, ai prompt widget now only calls onChange when the prompt content
changed (to not make useless calls to it)

Task-4780975

closes odoo/enterprise#90916

X-original-commit: c6dbe4587730e8d1f2f3f5417a7fa09117e66c7f
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Adrien Schoffeniels (adsc) <adsc@odoo.com>