[IMP] ai: show placeholder in prompt even when focused
Purpose:
--------
In the AI prompt widget, we want to show the placeholder even when the
cursor is inside the prompt but the prompt is empty (to handle the cases
where the prompt is autofocused).

Also added the Placeholder plugin that was missing.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>