[FIX] pos_restaurant_appointment: calendar event domain
Issue:
The calendar envents that have a restaurant table ressource are not
loaded in the pos if the config is in restaurant mode.

Fix:
If the config is a restaurant the calendar_event pos domain will return
a domain containing the appointment_ressources linked to the config
restaurant tables.

This commit also fixes the test that checks for the appointment-label on
the floor plan.

Task-id: 4970986
X-original-commit: 144a6d2

This commit adds hoot test for the pos_restaurant_appointment module.

closes odoo/enterprise#93806

X-original-commit: b1b4c8d
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Loukas Wets (lowe) <lowe@odoo.com>