[FIX] hide view signature request smart button when not needed
Before this commit, when logged in as a different user from the one responsible for counter-signing the contract, the smart button still showed, and clicking on it caused an error.

This commit fixes that, by hiding the smart button when there are no accessible signature requests.

Note that sign admins will still be able to see the smart button as they have access to the signature request.

Task-4966129

closes odoo/enterprise#93143

X-original-commit: 1daeca4e4720697b3af40e5b46fca0a3cd752840
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>