[FIX] l10n_co_edi_website_sale: handle obligation type codes with double digit ids
Problem: When there is an obligation type code with id greater than 9,
and it is selected in the dropdown of the website sale address form for
obligation type, the screen keeps loading forever and there is an
“expected singleton” traceback in the logs. This is because in the
method “_parse_form_data” in l10n_co_website_sale, the obligation type
field on form_data is set to be a  list of “type ids” which leads to an
error when convert_to_cache is called as the browse function in this
attempts to convert the list to a tuple of single characters. For
example, if the list is ["10"], it gets converted to ("1","0") hence
leading to the expected singleton traceback.

Purpose: Instead of passing a form list to form_data,we pass the record
set which will correctly set the values in the address, much like how
default_obligations_ids is also currently set. After this correction,
the website address screen will save the address properly and redirect
to the delivery screen for further actions.

Steps to Reproduce on Runbot:
1. Create a Colombian company, make sure l10n_co is installed
2. Set the company on the website to this company
3. Ensure that there is a record in the table
  l10n_co_edi_obligation_type_ids with id > 9. Create one if it does not
  exist.
4. Open the /shop page in incognito mode as a public user.
5. Add a product, go to the checkout page, proceed to the address page.
6. Enter all the information including the Identification Number
 (e.g. 623.456.789-1). Choose “NIT” in identification type and select
 the type code from step 3 in the dropdown for obligation type. Choose
 country “Colombia” along with a state and city
7.Click on "Continue checkout". The page gets stuck in a loading state
 forever.

opw-4776301

closes odoo/enterprise#96920

X-original-commit: 6df6fca0566a91f8b556cd46d73a1380975d6d88
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Nikita Desai (deni) <deni@odoo.com>