[IMP] ai: limit some fields size inserted in prompt
Purpose
=======
We can not prevent prompt injection if an admin wrote a prompt with
sensitive information, but we can try to make it more difficult
for someone to exploit and exfiltrate those information.

Limit the size of some fields (name, display name), for which it makes
no sens to have huge values.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>