[IMP] documents: load the actions asynchronously
Purpose
=======
On slow connection, when opening the cogwheel of the folder, there's
a lag. The reason is that we need to fetch the actions that can be
pinned on the folder, and we are waiting the response.

To fix that issue, we don't block `onWillStart`.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>