[IMP] documents: remove views sample
Support for this feature was only partial as when
we switched folder then came back to an empty one,
sample data was not rendered anymore.

The records are also rendered not muted during uploads,
which confuses users, and it's also tricky to avoid.

All things considered, we're removing support for the
sample records in Documents as it requires custom code
and is not really useful with the no-content helper.

Task-4788941

closes odoo/enterprise#90905

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>