[FIX] knowledge: indeterministic error with calendar event's resize
This commit fixes an indeterministic error in the
knowledge_calendar_command_tour when a calendar view is embedded and an
event is resized. The resize handler is only visible when hovering the
event, which is required to actually be able to resize it.

This commit merges both the "make resizer visible" step and the "resize
the item" step into a single one as those actions should be done in one
go, instead of separate steps (cf. the hovering state is kind of reset
in-between the steps).

Note: this matches the behavior of the CalendarView's unit tests helpers
(`resizeEventToTime()`).

runbot-error-163015

closes odoo/enterprise#91080

X-original-commit: fa28ed4ce820b29ca1aecdb15aaa4b8503c266dd
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>