[FIX] l10n_it_pos: change payment default index
In Italy if a payment method is of type 'Credit card' (type 2), and it's
index is 0 the payment made with it is considered as 'Not paid' by the
italian fiscal printer.

Why the fix:
------------
We change the default index of the payment method to 1, so that it is
considered as 'Paid' by the fiscal printer.

opw-4963421

closes odoo/enterprise#91273

X-original-commit: 411e7d2bc04edd77ee878175277c07e157199fdd
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>