[IMP] ai: don't send the full PDFs
Purpose
=======
Do not send the full PDFs to save cost. Instead, we do a heuristic,
and we send the start and the end of the PDFs.

We also set the image to low resolution in the same idea, it should
be enough for most cases.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>