[FIX] l10n_fr_reports: map box_26_external to JB in VAT return XML
The French VAT report line for grid 26 ("Repayment of credit requested on form n°3519")
uses code `box_26_external`, but the XML generator only mapped `box_26` to the `JB` tag.

As a result, the reimbursement amount was missing or incorrectly placed in the XML file.

This commit maps `box_26_external` to `JB` to ensure the correct tag is used
when the user fills in grid 26.

opw-4931275

closes odoo/enterprise#91488

X-original-commit: 5d0c9ed2ba7cfb6ec676ab464e8f5f1b1f4c06ce
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Oumaima Aarabe (ouaa) <ouaa@odoo.com>