[IMP] ai: fix fields and records protected in prompts
Purpose:
--------
Add FEFFs around the inserted records and fields in the AI prompt so
that pressing the enter key when the cursor is placed at the end of a
record/field element does not duplicate the element's span.

Also, the field and record selectors plugin now add the `oe-protected`
data attributes in `normalize`, so that even fields/records added from
outside of these plugins are protected.

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>