[IMP] ai: add Gemini support for AI server actions
Purpose
=======
Add support for Gemini in the AI server actions.

For embedding, transcription, etc, we use a "conversion to OpenAI layer",
but that layer does not support PDFs, and has some minor difference
with the OpenAI API that makes it annoying (different JSON schema,
we need to tweak the response to make it work, etc).

So, for the server action, we use the "standard API" (which supports
all the feature we want).

See
> https://discuss.ai.google.dev/t/combining-openai-compatible-gemini-completions-with-file-uploads/75281/2
> https://discuss.ai.google.dev/t/tool-calling-with-openai-api-not-working/60140

Task-4915266

Part-of: odoo/enterprise#90124
Related: odoo/odoo#218725
Related: odoo/upgrade#8073
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>