[IMP] l10n_be_hr_payroll_fleet: force car tracked value to be added in the employee chatter
As when signing the employee contract, the version related to that contract is
already created and we just set the version active, the value of the car does
not change, so the tracking is not triggered and it's not added into the
chatter.

This commit force the car_id and ordered_car_id to be tracked by adding the
initial value the employee precommit tracking values, before the version is
activated (so before the employee get the new values from the newly active
version).

Task-4962922

X-original-commit: f15016713d133258f4c22aae8f633a1695d3a1a8
Part-of: odoo/enterprise#96684
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>