[FIX] l10n_co_dian: not deleting DIAN email templates at module upgrade
Steps to reproduce:
- Install l10n_co_dian
=> "Invoice (DIAN): Sending" and "Credit Note (DIAN): Sending" email
templates are created
- Upgrade l10n_co_dian

Issue:
"Invoice (DIAN): Sending" and "Credit Note (DIAN): Sending" are deleted.

Cause:
Email templates are reset during a module upgrade.
These DIAN email templates are not defined in a XML, but created by a
python function.
This function is only called during installation.

Solution:
Set the "nopupdate" to True for these email templates to prevent them
to be reset during the module upgrade.

opw-4964347

closes odoo/enterprise#92453

X-original-commit: ce40f194d52fbab2c7ed105622519de284b5323f
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>