[FIX] knowledge: change to baseContainers config
This commit replaces the baseContainer config with baseContainers, a list that
defines the allowed base containers. The first element of the list now serves as
the default, equivalent to the previous baseContainer.

This change enables conditional display of "Normal" and "Paragraph" options in
the toolbar. For example, if baseContainers only includes "DIV", only the
"Normal" option will be shown. Similarly, if it only includes "P", only the
"Paragraph" option will be shown.

task-4770922

closes odoo/enterprise#91029

Related: odoo/odoo#216884
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>