[FIX] pos_urban_piper, pos_*: prevent duplicate preparation ticket
*: pos_urban_piper_enhancements

Before this change, when a device sent an order in preparation via the
ticket printer, this could result in the same order being printed by
multiple devices, as the order was not synchronized after it was sent.

The error is a bit tricky, because if the user had installed a
preparation screen, the order was sent to the preparation screen via
syncAllOrders. In this case, the order was correctly synchronized and
the other devices were informed of the changes.

This commit adds two things.
- We check the server before sending the order to preparation to make
sure it has not already been sent.
- Even when the user does not have a preparation display, the order
will be synchronized after being sent to a printer.

closes odoo/enterprise#91184

X-original-commit: 6cee9b45c5e221af95e8e514e2ebae56641f8876
Related: odoo/odoo#220919
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>