[FIX] website_sale_renting: hide list price from search engine
Versions
--------
- 17.0+

Steps
-----
1. Create a pricelist setting product prices to zero;
2. configure website to make zero-priced products unavailable for sale;
3. go to /shop as a public user;
4. open the product page of a zero-priced product;
5. open the browser console;
6. use `document.querySelector('.css_non_editable_mode_hidden .oe_currency_value')`.

Issue
-----
The query returns an element with the product's original list price.

Cause
-----
In `website_sale`, this element is part of a `t-if` that only gets
rendered when `editable` is `True`. A `website_sale_renting` override
changes the conditional to `not combination_info.get('is_rental')`.

Solution
--------
Instead of replacing the `t-if` attribute in the override, add
`and not combination_info.get('is_rental')` to it, ensuring it
remains hidden outside of editable mode.

opw-4865141

closes odoo/enterprise#91936

X-original-commit: e4a76466743113ffcb498e68c5300b19e2be6fa7
Signed-off-by: Levi Siuzdak <sile@odoo.com>