[FIX] stock_barcode: catch scan sound error
Issue
-----
On iOS 18.5, users can get a `NotSupportedError` when trying to scan a barcode
from the (barcode) main menu.

Steps to reproduce
-----
- Open barcode
- Click the center scan button
- Scan a barcode

--> Traceback

Discussion
-----
There are 2 issues occuring here.

1. We play either an ogg or mp3 file. However, the method to know if the format
is supported by the browser returns one of 'probably', 'maybe', ''.

https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement/canPlayType#return_value

This means that when we do

https://github.com/odoo/enterprise/blob/70e5013ea022ed04ff27db71ea3ecccb55ae1f70/stock_barcode/static/src/main_menu/main_menu.js#L37

We might end up using ogg even if the browser doesn't actually support it, as
'maybe' is truthy.

To reduce the risk of this happening, we can specify the codec as "vorbis" (all
ogg files of Odoo are vorbis).

2. The browser might not have the permission to play the sound. In such cases,
the best we can do is try to play the sound and log potential errors.

-----
Ticket:
opw-4820022

closes odoo/enterprise#91644

X-original-commit: f67447f4bb55c1bd5bf57e4a310f6a4f94ee8459
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: David Van Droogenbroeck (drod) <drod@odoo.com>