[IMP] l10n_be_reports{_hr_payroll}: annual account filing in XBRL
This commit adds the ability to export annual reports in XBRL
format to be able to file with NBB.

XBRL button added in Annual Statements report for Belgian companies
on the Belgian Annual Statements. The button opens a wizard for user
input of required data needed by the XBRL export.

New bridge module `l10n_be_reports_hr_payroll` added that depends
on `l10n_be_reports` and `l10n_be_hr_payroll` and installed automatically
when both are installed. The bridge module adds the social balance sheet
data needed for the XBRL export. If the module is not installed, then the
XBRL is generated without the social balance sheet information.

The XBRL is generated from template files. Each XBRL model has a different
template file. The XBRL model depends on the BS and PNL reports chosen in
the Annual Statement reports (con or cap & abbr or full) and the type of
the company (comp or assoc). Common tags are placed in a common template
to avoid duplication.

New fields (company region and company type) are added to the company. The
company region is computed automatically from the zip code and can be set
manually. The company type needs to be set manually before generating XBRL.

task-4866211

closes odoo/enterprise#93219

Signed-off-by: William André (wan) <wan@odoo.com>