[FIX] sale_renting_planning: sync planned dates with rental order dates
Before this commit, when the user updates the dates on rental order
linked to planning slots, the planning slots are not synchronised.
Same issue when a planning slot is rescheduled.

This commit makes sure the dates of rental order are recomputed once a
slot, linked to it, is rescheduled. By doing that, the dates of rental
order will always include the shifts.
When the dates of rental order are altered then the shifts could also
altered to make sure the shifts are still included in the rental period
of the rental order altered.

task-4916650

Part-of: odoo/enterprise#93670
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>