[REF] l10n_mx_edi_stock: Use fleet to handle Mx vehicles
Description of the issue/feature this PR addresses:
Users that have their own vehicles (the target user for carta porte also does) handle registries of their vehicles in fleet, but with the current implementation the user needs to create the vehicle in both carta porte and fleet making the purpose of fleet redundant.

Current behavior before PR:
The values for the vehicle in the bill of lading are taken from the l10n_mx.vehicle model, which causes users who have fleet records to have duplicate information.

Desired behavior after PR is merged:
The vehicle values on the bill of lading are taken from the fleet.vehicle model.
The vehicle driver (driver_id in fleet.vehicle) is taken as the intermediary (driver) on the bill of lading.
The computed method of the gross_vehicle_weight field has been reworked so that it is now always in tons.
Minor corrections were made to the bill of lading PDF.

opw-4117145

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#91185

Related: odoo/upgrade#8185
Signed-off-by: Laurent Smet (las) <las@odoo.com>