[FIX] sign: Delete unused sign item roles from the database
Fixes an issue where sign item roles were created but never deleted if they had no sign items linked to them. These orphaned roles were no longer shown in the app and remained in the database indefinitely.

To address this, a dummy sign item is now created for each new sign item role. This allows the autovacuum logic to detect and automatically clean up unused roles and dummy items. Without this dummy item, it would be impossible to determine if a role is truly orphaned.

task-4971485

closes odoo/enterprise#102953

X-original-commit: 4f5e33ada4b052c703e209cbd7ab7d216c34b717
Signed-off-by: Moustafa Mohamed Ibrahim Tawfik Aly (ibmo) <ibmo@odoo.com>