[FIX] whatsapp: markup formatting for italic in URLs requires whitespace
Steps to reproduce:
  1. Go to a WhatsApp template, for e.g. Sale Order.
  2. Update the sample value for the Portal Link variable to include an
  underscore.
  3. Click on Preview

Issue:
Due to the current regex, we will apply italic to any underscore we find
which is not ideal, mostly when we work with URLs that can contain
underscores like "acces_token" which we will try to apply italic to it
too.

Solution:
Following the most common logic for italic formatting, we will only
apply italic formatting when the underscore is surrounded by whitespace.
This is the same way it already behaves in Whatsapp.

If we write it like_this_for example, it will not apply italic
formatting, but if we write it like this _this_ it will apply italic
formatting.

So this for example:
 "If we write it like_this_for example"
 Will be like - before the fix:
  "If we write it like<i>this</i>for example"
 Will be like - after the fix:
  "If we write it like_this_for example"

opw-4720017

closes odoo/enterprise#92686

X-original-commit: 98e4a6e14a9bf8328e6d53dd224d9555f813ab80
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>