[FIX] account_accountant: rectify prediction of taxes
Previously, when _predicted_field was invoked from _predict_specific_taxes, self
did not have an associated move_id. As a result, _build_predictive_query was
unable to construct the correct query, causing the prediction logic to fail and
return False.

With this fix, a new argument has been added to _predicted_field to pass the
move. This ensures that the prediction logic uses the correct move_id and
behaves as expected.

task-4943015

closes odoo/enterprise#91413

X-original-commit: 108bfbf6ccc410ba7c342f93fa3d94b3ac924b45
Related: odoo/odoo#221330
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Harsh Trivedi (hatr) <hatr@odoo.com>